refractor bad fetch loop into single component
