+++ src/components/studentList.js
@@ -1,15 +1,10 @@
 import React from "react"
 // import { useState, useEffect } from "react"
 
-import Students from "../utils/students"
-
 const StudentList = () => {
-  const data = Students()
-  if (!data) return <p>Loading...</p>
-
   return (
     <div>
-      <h1>this will be the student list table</h1>
+      <p>this will be the student list table</p>
     </div>
   )
 }
